requirements.txt
requests==2.27.1
beautifulsoup4>=4.8.1
gspread>=5.5.0
oauth2client>=4.1.3
Flask==2.2.2
python-dotenv==0.21.0
gunicorn==20.1.0
